Clean up version test and add #ifndef checks for macros
!(nfhdr->nflog_version) is equivalent to (nfhdr->nflog_version == 0).

That will evaluate to 1 if nfhdr->nflog_version is 0 and to 0 otherwise.

So !(nfhdr->nflog_version) == 0 is equivalent to nfhdr->nflog_version != 0,
but 1) it's more obvious what it means and 2) compilers don't point out
that it may not mean what you intended.

This is a cherry-pick of upstream commit:
https://github.com/the-tcpdump-group/libpcap/commit/f94cd1eb7d0ab32bf1db1dd029acc0c68b054363

Additionally, add #ifndef checks for BPF_STMT and BPF_JUMP macros.

Bug: 26936282
Change-Id: Id5433d1dcb8b866f7b358174962add447801c9d8