Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > +hansson/dfreni for the idea that there should be some way to say in the build system "i want the oldest API level that modules need to support", rather than "current" which will be wrong next year, or hard-coding numbers that lose the intention-revealing quality of something like `sdk_version: "min_apex"` or whatever.
> 
> Another strategy would be to pick the lowest possible number, ignoring apex. e.g. protobuf uses 9 (lowest finalized NDK?)

well, please never hard-code any number. 3 was the first NDK, 16 is the current lowest API level supported by the NDK ... and this "9" is a perfect example of the problem here. who knows how/why they chose that and what their intention was?

(specifically, if your intention is "maximum compatibility", you should use `sdk_version: "minimum"`. that will just slowly march forward with time without us having to touch hundreds of .bp files.)

> Is there a reason to build against something newer than that for native? For APKs it might not bethe best fit because I think the sdk_version can end up in the targetSdkVersion attribute which can affect runtime behavior.

until 21, the C library was in pretty bad shape. lots of code needs Android-specific hacks up to 21. past 21, it's mostly fine. but if you do see someone use a random number > 21 it's often because they need some function that wasn't added until 24 or whatever. (in this specific case, libpcap code needs getifaddrs() which wasn't in bionic until -- coincidentally because i had to look it up -- 24.)

but that's the point: you can't tell anything from a number. "current" and "minimum" are both great, but i feel like there's a gap opening up if we now have folks needing "the lowest API i need to still support because of modules", which -- like current and minimum -- actually changes over time.




(to be clear: all i'm asking for from lorenzo is a code comment and a better checkin message. but i think longer term there needs to be a more intention-revealing way to say "this code ends up in a module, so use the right SDK version" in the build system. or if that already exists, that's what should be used here.)

Patch-set: 4
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>